Extract buildInfo from parent pipelines
Before this orca can extract artifacts from triggering pipelines of depth 1.
This code enables orca to extract artifacts from pipelines in infinite depth
and also extract the build details from the build that produced the artifact.
The objective of this PR is to populate buildInfo in amis created with artifacts
from parent pipelines.