Handle an Execution failing during the startup process.
Right now there is a case where an execution is persisted in the initial
NOT_STARTED state, before the stage types and configurations are evaluated.

If that evaluation process fails (for example a stage that uses one of the
deprecated TargetServerGroup types), that execution never starts but remains
in the NOT_STARTED state, and we don't surface that failure to the user.

This adds an exception handler around startup, and if there is a failure
during the initial startup phase, the execution is set to TERMINAL, and
marked as cancelled by the 'system' user with a cancellation reason showing
the failure cause.