feat(rollback): Support rolling back to a server group that no longer exists
This PR provides a rollback strategy that will clone forward with the
image details that were tagged as part of spinnaker/orca#1705.

It also supports the `imageName` being explicitly provided, but that's
an exceptional case that would not be supported in the UI.

The `spinnaker:metadata` tag will include `buildInfo` for the image if
it is returned from `clouddriver`.