feat(core): Support for rolling back a cluster (across multiple regions)
This PR is motivated by a desire for a smarter
rollback capable of operating on more than one
region.

It builds on top of the existing server group
rollback capabilities and makes a determination
on what to rollback based on a two factors.

Whether or not the current server group has a
a `spinnaker:metadata` entity tag indicating
the previously deployed image.

If the entity tag exists, we will attempt to (a)
enable an existing server group with that image
or (b) perform a red/black with the previous image.

If the entity tag does not exist, we will attempt
to rollback to the most recent previous server
group, regardless of what image it was deployed
with.

This is intended to be cross cloud provider but
your mileage may vary with anything beyond AWS
right now.