fix(webhook): Deprecate use of buildInfo
The webhook stage puts a lot of arbitrary information into buildInfo, that broke assumptions done other places in Orca.
This commit deprecates the use of buildInfo, and adds information about the deprecation and how to migrate, to the stage output. The webhook output is now added beneath the key `webhook`.
The deprecated stuff should be removed after some time.