feat(aws): Support for pinning min=desired capacity on source server group
This PR attempts to avoid unnecessary and problematic autoscale events
that occur as the total # of instances increase when a deploy is
happening.

As instance count increases, there is natural downward pressure
on the cluster resulting in the source server group being scaled down.

It is particularly bad for rolling red/black where the old server group
is not immediately disabled.

The expectation is that the source group will be returned to its
original unpinned capacity when the deploy completes (or fails!).