fix(after stages): Allow for just-in-time planning of after stages
This is a bridge to the 6.5.x branch of Orca where after stages are _always_ planned just in time. During deployment this can be an issue if a 6.5.x instance starts a stage (and does not plan after stages) and a 6.4.x instance tries to complete the stage (finding no before stages or tasks and no planned after stages).