config(metrics): record task completion time
This changes the meaning of the task.invocations.duration metrics to
record only the time taken by this particular invocation, not the time
taken since the beginning of that execution.

In CompleteTaskHandler, we record the entire time it took for the
execution to complete.

In the example below showing the lifecycle of a task:

StartTask  RunTask    RunTask            RunTask         CompleteTask
 |----------|---|-------|-|--------------|-----|---------------|--> time
            ^-A-^       ^B^              ^--C--^
 ^-----------------------------D-------------------------------^

We would have:
* task.invocations.duration recording span A, B and C (3, 1 and 5 ticks respectively)
* task.completions.duration recording span D

Additionally, this also removes the counters, as we get them for free
with the timers (no need to double count)