fix(clouddriver): Handle situations where the wrong `pinned_capacity` tags were deleted
Only check for entity tags if the current stage context contains
sufficient coordinates to identify the target server group.

```
@GET('/tags')
List<Map> getEntityTags(@QueryMap Map parameters)
```

If `parameters` contains a null value (say for an account or server group
name), it will be stripped out and not sent across the wire.

This effectively changes a query for a very specific server group and
account into a wildcard.

The best way to ensure that server groups are correctly being unpinned
is to enable the `RestorePinnedServerGroupsPoller` is enabled.

`pollers.restorePinnedServerGroups.enabled: true`