fix(orca) fix force cache refresh sometimes taking 12 minutes (#3366) (#2422) (#2424)
Sometimes the "Force Cache Refresh" task of deploying a v2 manifest would take
up to 12 minutes to complete. This would occur as result of a timing issue
between when Orca submits the refresh requests to the cloud driver and when the
cloud driver processes the requests.

The fix is to handle the scenrio where Orca "forgets" pending requests that
have been previously completed if not all the pending requests are completed
when it queries the cloud driver.