fix(pipelines): make evaluation context and transform public
Currently one has to use PipelineExpressionEvaluator.evaluate outside of this package, but that has a lot of limitations. It might be useful to keep an evaluationContext around and run multiple transforms against it or use a different ExpressionParser. I see PipelineExpressionEvaluator as an example of how expressions can be evaluated, but not an abstraction that shouldn’t be opened. Allowing context and transform creation from outside the package makes sense.