fix(FindImageFromCluster): infer regions from deploy stages
FindImageFromClusterTask now looks ahead to find all regions from subsequent deploy (regular + canary)
stages. This moves the failure point of a pipeline up when an image isn't found.
E.g. FindImage specifies to look for images in `us-east`, but a deploy right after was modified to
deploy to `us-east` AND `us-west`. Thuse FindImage will locate the `us-east` image,
the deploy for that image will start but the deploy for `us-west` will fail.
This failure can cause regions to have different versions deployed.
To prevent this, we now will fail in the FindImage task before any deploy starts

Additionally, updated the BakeStage to take advantage of this functionality