feat(clouddriver): Consider target capacity if supplied from `clouddriver`
AWS deployments are somewhat special in that they create an initial
asg of size 0/0/0 and then subsequently resize it to the target
capacity.

This PR protects against premature success by favoring the
clouddriver-supplied capacity if we see a 0/0/0 asg but expect a
non-0/0/0 asg.

The edge case only appears when the `ForceCacheRefresh` task is dropped
from the deploy stages in `orca`.

relates to spinnaker/clouddriver#3356