fix(securityGroups): User `securityGroupName` for upsertSecurityGroup tasks
The upsert task need the `name` parameter from context, which is removed by orca because it "knows about it"
which makes it impossible to run this task inside a pipeline (but ok as a job/task).
The name is needed inside clouddriver so it can't be `null`. Instead add a new property `securityGroupName`
that contains the name which won't be munged out of the context.

`deck` changes are forthcoming, for now, this is a backwards compatible change