fix(execution): a stage with redirect tasks gets marked terminal
The logic for determining stage status seems overly strict.
The redirect execution status is used in a task loop. Without counting that as success how would something like RollingPushStage ever succeed? Or do the redirect statuses get modified somehow?
I could also make an argument that skipped tasks should not kick the stage out of opportunity to be a success.