fix(kubernetes): Don't fail deploy for SpEL errors in manifest artifacts
Fixes spinnaker/spinnaker#3644
This will cause a warning to be displayed in the pipeline UI, just
as when an evaluation error happens in a manifest in the stage as text.
It allows the successful deployment of manifests that contain expressions
that are evaluated by the pod (e.g. configmaps with Spring config)