test(ci): Add tests to JenkinsStage and ScriptStage
There are no tests that waitForCompletion is properly read when
starting a Jenkins stage; this functionality recently broke, so
add some tests to it.  Also add some tests to verify that binding
of artifacts works correctly.

The ScriptStage has no tests at all; add a simple test that it
fetches the properties after the script finishes.