fix(artifacts): Fix successful filter for find artifacts
When filtering to only include successful executions, the
front-end sets the flag 'successful' in the stage config.
Orca is incorrectly looking for a value 'succeeded', and
thus ignores the value that was set on the front-end.

While 'succeeded' is slightly better named as it matches the
execution status, changing this on the backend means that
existing pipelines will work automatically.  (Whereas changing
it on the front-end would require either a pipeline migrator
or for anyone affected to re-configure the stage.)