fix(artifacts): Fix successful filter for find artifacts (#2780) (#2782)
When filtering to only include successful executions, the
front-end sets the flag 'successful' in the stage config.
Orca is incorrectly looking for a value 'succeeded', and
thus ignores the value that was set on the front-end.

While 'succeeded' is slightly better named as it matches the
execution status, changing this on the backend means that
existing pipelines will work automatically.  (Whereas changing
it on the front-end would require either a pipeline migrator
or for anyone affected to re-configure the stage.)

To avoid breaking anyone who manually edited their stage config to
set 'succeeded' because that's what the back-end was looking for,
continue to accept that in the stage config.