defer RUNNING status until after the extra initial store.
Being in the RUNNING state affected the handler guard checking
for NOT_STARTED (line 71) causing it to be ignored. I dont
really understand the control flow here, and could not follow
along in a debugger because kotlin/mockito jumps all over the
place making it impossible for me to grok that way. I suspect
there is a better fix and could involve adding another status
between NOT_STARTED and RUNNING but that is beyond my level
of intimacy with the model here.

fixed tests to accomodate the second storeStage call.