refactor(gcb): Use generic maps for GCB objects
Orca currently deserializes the build configuration in the stage to
a Build object only to send it immediately over to igor (which
re-serializes it). As orca doesn't actually need to know any of the
details about the Build object it's sending over, just use a generic
Map.

This addresses an issue where some fields are not correctly
deserializing using the default objectMapper; I'll need to
work around this in igor, but this reduces the scope of where
that workaround needs to live.  (This also allows us to
better encapsulate which microservices actually need to know
about GCB objects.)