refactor(provider/kubernetes): Replace nested maps with an object
The ManifestForceCacheRefreshTask currently keeps track of its manifests
as a Map<String, List<String>> which leads to a lot of complex iteration
over this structure. Create a new class ScopedManifest and flatten this
structure into a List<ScopedManifest> so that it's much easier to
follow the processing that the class is doing.