refactor(provider/kubernetes): Add account to ScopedManifest
Rather than thread account through multiple calls in this class, add
it to the ScopedManifest class and set the account on each manifest
when we create the initial list.

This also makes the helper class Details identical to ScopedManifest,
so replace instances of Details with ScopedManifest.

Finally, replace pendingRefreshProcessed which both returns a status
and mutates the stage context with getRefreshStatus, which only
returns a status. Leave mutation of the context to the caller,
checkPendingRefreshes. This allows us to avoid needing to mutate
a ScopedManifest and keep the class immutable.