refactor(provider/kubernetes): Change control flow in refresh
Instead of having refreshManifests and checkPendingRefreshes
contain logic for determinig if the task is done, have them
focus on mutating refreshedManifests and deployedManifests
as they take actions.

Then add allManifestsProcessed to check whether we're done; which
is just checking if all deployed manifests have been processed.
This removes the need to track the state of all manifests in the
mutating functions and allows us to consolidate the return value
of the task to one place.