fix(provider/kubernetes): Don't poll immediately after cache refresh
We currently poll clouddriver to get the status of a cache refresh
immediately after requesting the cache refresh, and schedule a
re-refresh is we don't see the request we just requested.

For users with a read-only clouddriver pointed at a replica of redis,
there will be some replication lag before the pending refresh
appears in the cache, which will cause us to keep re-scheduling the
same cache refresh.

To address this, wait one polling cycle before checking on the status of
a pending refresh. Do this by changing the order of operations so that
we first check on any pending requests from the last cycle, then
schedule and needed new requests.