refactor(ci): Generify RetryableIgorTask
I'd like to re-use the logic in RetryableIgorTask for some GCB
tasks, but it requires that the stage map to a CIStageDefinition.
Make the stage definition a parameter to the class so it can be
re-used.

At some point it might make sense to make this even more general than
just for igor tasks, but at least this makes it a bit more general
than it is now.