fix(artifacts): Find from execution should use default artifact
The public method ArtifactResolver.resolveSingleArtifact should
handle trying to match the artifact in the context as well as
falling back to a prior execution or the default artifact instead
of making callers do that.

Pull the matching logic into the private function matchSingleArtifact
then have resolveSingleArtifact call matchSingleArtifact and
then do the defaulting if the artifact isn't resolved. We can pull
this logic out of resolveExpectedArtifacts and have it just call
resolveSingleArtifact.