perf(artifacts): only load prior artifacts when needed
When a pipeline is triggered via `POST /orchestrate`,
OperationsController will resolve artifacts for the pipeline. The
behavior of `ArtifactResolver.resolveArtifacts(Map)` will load all
previous executions of the pipeline in order to find prior artifacts of
the pipeline (from the most recent execution), even if the
`priorArtifacts` list is never accessed (because the pipeline trigger
contains all expected artifacts).

This change tries to avoid the unnecessary loading of previous
executions / prior artifacts by wrapping the calls to
getArtifactsForPipelineId() etc in a `Supplier` so that the data (which
could be voluminous if old pipeline execution cleanup is not enabled) is
only loaded when truly needed.

To make the change a little bit smaller, I refactored the signature for
the public `resolveSingleArtifact` method to remove the `priorArtifacts`
parameter which was always null in all call sites.

See https://github.com/spinnaker/spinnaker/issues/4367 for more details.