fix(pipeline cleanup): remove 2nd check on thresholdDays
Fixes https://github.com/spinnaker/spinnaker/issues/4467

OldPipelineCleanupPollingNotificationAgent is checking the age of the
pipeline execution twice before removing it:

1. the `Observable<Execution>` returned from
`executionRepository.retrievePipelinesForApplication(app)` is filtered
to only executions that have a completed status and where startTime (or
buildTime, if startTime is not set) is before `now - thresholdDays days`

2. After converting the `Observable<Execution>` to a `List`, sorting it,
and retaining `minimumPipelineExecutions` elements, each execution is
only deleted if `days between startTime and now > thresholdDays`.

The first check is true for executions as soon as they are a moment
older than `executionDays` days, but the second requires executions to be
older than `executionDays + 1` days (due to using `>`).

Since the second check seems redundant, I've simplify removed it.

My assumption here is that it is a bug that a configuration setting like
`thresholdDays: N` actually only removes executions older than `N+1`
days.